# pets-orders-openapi/src/main/resources/openapi/pets-orders-openapi.yaml
openapi: 3.0.3
info:
  title: Pets & Orders API
  description: API for managing pets and pet orders
  version: 1.0.0

servers:
  - url: http://localhost:8080

tags:
  - name: Pets
    description: Operations related to pets
  - name: Orders
    description: Operations related to orders

paths:
  /pets:
    get:
      tags:
        - Pets
      summary: List all pets
      operationId: listPets
      parameters:
        - name: limit
          in: query
          description: Maximum number of pets to return
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: A paged array of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"

    post:
      tags:
        - Pets
      summary: Create a new pet
      operationId: createPet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPet"
      responses:
        "201":
          description: Pet created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        "400":
          description: Invalid input

  /pets/{petId}:
    get:
      tags:
        - Pets
      summary: Get a pet by ID
      operationId: getPetById
      parameters:
        - name: petId
          in: path
          description: ID of the pet
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: A single pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        "404":
          description: Pet not found

    delete:
      tags:
        - Pets
      summary: Delete a pet by ID
      operationId: deletePet
      parameters:
        - name: petId
          in: path
          description: ID of the pet
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Pet deleted successfully

  /orders:
    get:
      tags:
        - Orders
      summary: List all orders
      operationId: listOrders
      responses:
        "200":
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      tags:
        - Orders
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewOrder"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input

  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get an order by ID
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          description: ID of the order
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: A single order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

    delete:
      tags:
        - Orders
      summary: Delete a order by ID
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          description: ID of the order
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Order deleted successfully

components:
  schemas:
    Pet:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Fluffy
        tag:
          type: string
          example: cat
        type:
          type: string
          example: pet-animal

    NewPet:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Fluffy
        tag:
          type: string
          example: cat
        type:
          type: string
          example: pet-animal

    Order:
      type: object
      required:
        - id
        - petId
        - quantity
        - status
      properties:
        id:
          type: integer
          format: int64
          example: 1001
        petId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int32
          example: 2
        status:
          type: string
          example: CREATED

    NewOrder:
      type: object
      required:
        - petId
        - quantity
      properties:
        petId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int32
          example: 2